fix table count + re-add depth check in DataTable::__destruct + reset tables in pluginsarchiver as in archiveprocessor

+++ core/DataTable.php
@@ -223,7 +223,7 @@ class DataTable implements DataTableInterface, \IteratorAggregate, \ArrayAccess
     /**
      * Maximum nesting level.
      */
-    public static $maximumDepthLevelAllowed = self::MAX_DEPTH_DEFAULT;
+    private static $maximumDepthLevelAllowed = self::MAX_DEPTH_DEFAULT;
 
     /**
      * Array of Row
@@ -354,7 +354,12 @@ class DataTable implements DataTableInterface, \IteratorAggregate, \ArrayAccess
      */
     public function __destruct()
     {
-        if (isset($this->rows)) {
+        static $depth = 0;
+        // destruct can be called several times
+        if ($depth < self::$maximumDepthLevelAllowed
+            && isset($this->rows)
+        ) {
+            $depth++;
             foreach ($this->rows as $row) {
                 Common::destroy($row);
             }
@@ -363,6 +368,7 @@ class DataTable implements DataTableInterface, \IteratorAggregate, \ArrayAccess
             }
             unset($this->rows);
             Manager::getInstance()->setTableDeleted($this->currentId);
+            $depth--;
         }
     }
